// generated/src/app/api-models/stage-claim-ndc.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class StageClaimNdc {
  batchId: string;
  transactionId: number;
  lineNo: number;
  subLineCode: string;
  sequenceNo: number;
  ndcCode: string;
  quantity: number;
  codeQualifier: string;
  ndcAmount: number;
  ndcUnitMeasure: string;
  ndcRxCodeQualifier: string;
  ndcRxNumber: string;
  insertDatetime: Date;
}